requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
python-multipart==0.0.12
tensorflow==2.15.0
numpy>=1.26.0,<2.0.0
Pillow==10.4.0
pydantic==2.9.2
protobuf==3.20.3
h5py>=3.11.0
